Move dispatching types to own module.
