+++ http2-client.cabal
@@ -20,6 +20,7 @@ library
                      , Network.HTTP2.Client.Helpers
                      , Network.HTTP2.Client.RawConnection
   other-modules:       Network.HTTP2.Client.Channels
+                     , Network.HTTP2.Client.Dispatch
   build-depends:       base >= 4.7 && < 5
                      , async
                      , bytestring
